+++ CHANGELOG.md
@@ -13,6 +13,7 @@ A more detailed list of changes is available in the corresponding milestones for
   - **[com.google.fonts/check/metadata/family_directory_name]:** We want the directory name of a font family to be predictable and directly derived from the family name, all lowercased and removing spaces. (issue #3421)
   - **[com.google.fonts/check/file_size]:** Ensure that the absolute file size of the font is not going to cause problems on the Google Fonts platform. (issue #3320)
   - **[com.google.fonts/check/render_own_name]:** Ensure the font can render its own name without .notdef glyphs. (issue #3159)
+  - **[com.google.fonts/check/varfont/grad_no_reflow]:** Ensure that variations on the GRAD axis do not change any horizontal advances. (issue #3187)
 
 
 ## 0.8.1 (2021-Aug-11)